backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
supabase
pinecone-client
langchain
langchain-core
langchain-pinecone
langchain-google-genai
langchain-groq
python-multipart
python-dotenv
